Renamed "TRR test suite" to "TRR verification package", shout out to Michael!
