+++ src/nectarchain/acceptance_verification_package/gui.py
@@ -3,8 +3,8 @@ import sys
 from PyQt5.QtWidgets import QApplication
 
 from nectarchain.acceptance_verification_package import hillas_validation, ped_vs_time
-from nectarchain.trr_test_suite import deadtime
-from nectarchain.trr_test_suite.gui import TestRunner
+from nectarchain.trr_verification_package import deadtime
+from nectarchain.trr_verification_package.gui import TestRunner
 
 
 class AcceptanceTestRunner(TestRunner):
