+++ src/nectarchain/trr_verification_package/resources/bad_pix_module.json
