+++ src/nectarchain/user_scripts/hashkar/final_plotting/pix_tim_plot.py
@@ -5,7 +5,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 
-from nectarchain.trr_test_suite.utils import pe2photons, photons2pe
+from nectarchain.trr_verification_package.utils import pe2photons, photons2pe
 
 plt.style.use("../../../utils/plot_style.mpltstyle")
